Package qemu
On macOS, this builds qemu with homebrew, based on upstream formula but
limited to just what we need.  This is so we can reuse finch's script to
identify the required files.

On Linux, we build qemu directly, and use the scripts from the old
lima-and-qemu repo to identify the files we need to ship.

Because the aarch64 macOS runners on GitHub do not support
virtualization, we hack in a few extra files to include on that
platform.  The list of files is based on what the x86_64 run had.

We need slirp because lima uses that (with `--netdev user,...`).

Signed-off-by: Mark Yen <mark.yen@suse.com>