Linux: Exclude more glib2 depedencies
It looks like our list was ultimately derived from an upstream AppImage
list, but that list has in the mean time changed to start bundling GLib.
However, we must not bundle GLib as Electron also depends on it, and
shipping them would result in Electron failing to start up.  Instead, we
go the other way and exclud more GLib dependencies (in particular,
libmount.so).

Signed-off-by: Mark Yen <mark.yen@suse.com>